requirements-dev.txt
# Include the production dependencies
-r requirements.txt

pytest-django==3.1.2
factory_boy==2.9.2
